requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
docopt==0.6.2
findspark==2.0.1
greenlet==1.1.3
hdfs==2.7.0
idna==3.4
kafka-python==2.0.2
mrjob==0.7.4
numpy==1.23.3
pandas==1.5.0
psycopg2-binary==2.9.4
py4j==0.10.9.5
pyspark==3.3.0
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
urllib3==1.26.12
